refactery: move mem_engine/tests to traits/test_base
